Dinomark plugin (#3)
the Dinomark plugin extends static markdown with dynamic super powers. it exploits block-level `[linkref]: link (comment)` and inline links `[text][ref]` and `[text](link)` in specific ways so that non-compliant processors will simply ignore the markup.

the following example demonstrates usage of out-of-the-box support:

```
# sets `ctx.vars.var.name` with `"string"`
[@var]: var.name (string)
# sets `ctx.vars.var.name` with json-decoded value of `json-string`
[@var]: var.name (json:json-string)

# reads file (json) and puts value into `ctx.vars.var.key`
[@include-var]: path/to/properties (var.key)

# reads file and includes content as-is
[@include] path/to/file

# reads file and sends through document processor for formatting
[@process] path/to/file
```

## link reference blocks

```
[@directive]:
[@directive]: action
[@directive]: action (parameters)
```

executes a `directive` handler. multiple directives can be used in a block. directives are extensible.

## link tags

```
use [][var.reference] or [](statement) to execute dynamic inline processing
```

note the `[]` -- this ensures the link output is suppressed for non-compliant processors