Data Registry (#7)
the data registry allows handlers to publish pieces of info (grouped by a category) that other handlers can use. the first use case is in HeaderHandler, which publishes the header content, id, and level for each new header it encounters. this can enable things like auto table of contents or section navigation